Use 16bit unsigned for channel id
Bug: none

+++ p2p/base/turn_port.h
@@ -192,7 +192,8 @@ class TurnPort : public Port {
   }
   // Finds the turn entry with `address` and sets its channel id.
   // Returns true if the entry is found.
-  bool SetEntryChannelId(const rtc::SocketAddress& address, int channel_id);
+  bool SetEntryChannelId(const rtc::SocketAddress& address,
+                         uint16_t channel_id);
 
   void HandleConnectionDestroyed(Connection* conn) override;
 
@@ -304,7 +305,7 @@ class TurnPort : public Port {
   void HandleDataIndication(const char* data,
                             size_t size,
                             int64_t packet_time_us);
-  void HandleChannelData(int channel_id,
+  void HandleChannelData(uint16_t channel_id,
                          const char* data,
                          size_t size,
                          int64_t packet_time_us);
@@ -323,7 +324,7 @@ class TurnPort : public Port {
 
   bool HasPermission(const rtc::IPAddress& ipaddr) const;
   TurnEntry* FindEntry(const rtc::SocketAddress& address) const;
-  TurnEntry* FindEntry(int channel_id) const;
+  TurnEntry* FindEntry(uint16_t channel_id) const;
 
   // Marks the connection with remote address `address` failed and
   // pruned (a.k.a. write-timed-out). Returns true if a connection is found.
